Remove range check for luaTicksLimit & eventListenerLuaTicksLimit - all longs are allowed

+++ src/main/java/net/wizardsoflua/lua/SpellProgram.java
@@ -87,9 +87,9 @@ public class SpellProgram {
 
     Clock getClock();
 
-    int getLuaTicksLimit();
+    long getLuaTicksLimit();
 
-    int getEventListenerLuaTicksLimit();
+    long getEventListenerLuaTicksLimit();
 
     boolean isScriptGatewayEnabled();
